+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add some snippets for autocompletion i.e. `WHERE`, `JOIN`, `ON`, `AS`, `AND`, `OR`
 - Add subquery autocompletion
 - Add a loading badge for attribute autocompletion
+- Support impEx (Execution, Validation)
 
 ## 0.0.6